pyproject.toml
[project]
name = "ar-classification"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "hydra-core>=1.3.2",
    "hydra-ray-launcher>=1.2.1",
    "ipykernel>=7.0.1",
    "matplotlib>=3.10.7",
    "numpy>=2.3.4",
    "pandas>=2.3.3",
    "ray>=2.50.1",
    "seaborn>=0.13.2",
    "torch>=2.9.0",
    "tqdm>=4.67.1",
    "wandb>=0.22.2",
]

[[tool.uv.index]]
name = "pytorch"
url = "https://download.pytorch.org/whl/cu126"
explicit = true

[tool.uv.sources]
torch = { index = "pytorch", marker = "sys_platform == 'linux'" }
torchvision = { index = "pytorch", marker = "sys_platform == 'linux'" }
torchaudio = { index = "pytorch", marker = "sys_platform == 'linux'" }
